Working on double large prime variation - work in progress
Updated the Hashmap functions to correct bugs and be closer to needs.
Updated the initialization of PartialRelations instances.
Updated the relations.c file to adapt to the new data structure used.
For now, I work only on one large prime, but with the new data structure. Now I will work on the functions needed to build the graph and find cycle in the graph. Smoothness tests will be updated to return the two large primes (commented now).